rename files and deprecate slope_opentopo
