+++ tests/test_tm_utils.py
@@ -5,7 +5,7 @@ from gri_shared_library.os_tools import create_folder
 
 from conftest import DT_TEST_PARAMS_DIR, SECRETS_FILE_PATH, TEST_01_GRI_PROJECT_ID, TEST_REAL_PROJECT_C1_ID
 from decision_tree.api_utils import download_geoparquet
-from decision_tree.slope import opentopo_pull_wrapper
+from decision_tree.legacy.slope_opentopo import opentopo_pull_wrapper
 from decision_tree.process_api_results import _read_geoparquet, flatten_tm_geoparquet, TestProjectHandling
 from decision_tree.process_api_results import process_tm_results
 from decision_tree.tools import convert_to_os_path, load_secrets, load_yaml
